Reduce tx cache memory footprint by half
Trim the transaction keys by half before using them as cache key. The
rationale is that even at half the size for the purposes of short lived
caching the chances of collition is astronomically unlikely. This then
allows us to use double the cache size with the same memory footprint.

As part of the same work reduce the number of calls to tx.Key to avoid
recomputing the sha256 hash of the raw transaction payload. sha-256 is
a relatively expensive hash function. Therefore, reducing unnecessary
calls to key would reduce the compute footprint of the mempool.
Specifically, refactor mempool tx cache to take tx key directly instead
of raw tx itself.

Note that List receiver is now removed from the LRU cache implementation
to accommodate the fact that keys are trimmed. This function was never
used.

Refactored TTL duplicate tx cache to also use a trimmed key.