Cache the last asked for percentile for 5 seconds
Cache the last asked for percentile for at least 5 seconds and avoid
redundant re-computation of it if the sample set has not changed. This
cache design is based on the rationale that the percentile value asked
for in Sei is fixed in configuration for the lifetime of the
application. Therefore, the only value that's updated in reservoir is
the samples.