Fixed timeoutTicker (#331)
timeoutTicker should be a long running task. Due to a bug (context
shadowing), the task was terminating early if all the following
conditions occurred at the same time:
* timeout happened
* output channel is full (task reporting the timeout is blocked)
* new timeout arrived on the input channel


Fix simplifies the timeout management by spawning 2 long running tasks
(instead of task per timeout). Additionally:
* A test was added specifically to detect early exit of the ticker task.
* timeoutTicker task is now spawned via Service.SpawnCritical (instead
of Service.Spawn), which panics in case the task fails instead of just
logging an error.
* Service.SpawnCritical now panics in case the spawned task terminates
with context.Canceled before the service is canceled (it used to be
silently ignored, equivalent to a task success). This is a temporary
measure until the Service is replaced by structured concurrency
altogether.